업데이트 완료는 뜨지만 수정이 안됨.

+++ com.mire.cinema/src/main/java/com/mire/cinema/domain/movieschedule/MovieScheduleDTO.java
@@ -35,7 +35,7 @@ public class MovieScheduleDTO {
 		private LocalDateTime scheduleFinishTime;
 		private long screenNo;
 		private long movieNo;
-		private int moviePlayTime;
+		private float moviePlayTime;
 	}
 
 	@Data
